prommon/app.py
__copyright__ = "(c) 2018 RenaissanceRe IP Holdings Ltd.  All rights reserved."

import signal
import argparse
import logging

from tornado import ioloop, web, gen
from tornado.log import app_log

from prometheus_client import Counter, generate_latest, CONTENT_TYPE_LATEST
